+++ src/main/java/elegant/children/catchculture/service/visitAuth/VisitAuthService.java
@@ -11,6 +11,7 @@ import elegant.children.catchculture.entity.pointhistory.PointChange;
 import elegant.children.catchculture.entity.user.User;
 import elegant.children.catchculture.entity.visitauth.VisitAuth;
 import elegant.children.catchculture.event.CreatePointHistoryEvent;
+import elegant.children.catchculture.event.DeleteFileEvent;
 import elegant.children.catchculture.event.SignOutEvent;
 import elegant.children.catchculture.repository.culturalEvent.CulturalEventRepository;
 import elegant.children.catchculture.repository.visiatAuth.VisitAuthQueryRepository;
@@ -90,8 +91,10 @@ public class VisitAuthService {
                 .flatMap(Collection::stream)
                 .map(fileUrl -> FileEvent.builder().fileName(fileUrl).build())
                 .collect(Collectors.toList());
-
-        applicationEventPublisher.publishEvent(fileEvents);
+        if(fileEvents.isEmpty()){
+            return;
+        }
+        applicationEventPublisher.publishEvent(new DeleteFileEvent(fileEvents));
         visitAuthRepository.deleteByUserId(signOutEvent.getUserId());
     }
